v3.8.0 update (#2082)
* 3.8 update

* fix Markus' name

---------

+++ include/cutlass/detail/collective/mixed_input_utils.hpp
@@ -301,7 +301,7 @@ struct LayoutAwareConvertImpl<
     }
   }
 };
-
+/*
 // Specialization for E5M2 -> FP16 with [3120] value order
 template <>
 struct LayoutAwareConvertImpl<
@@ -343,12 +343,12 @@ struct LayoutAwareConvertImpl<
     }
   }
 };
-
+*/
 // Specialization for INT8 -> BF16 with [3120] value order
 template <>
 struct LayoutAwareConvertImpl<
   cutlass::int8_t,
-  cutlass::half_t,
+  cutlass::bfloat16_t,
   cute::Layout<cute::Shape<_2,_2>, cute::Stride<_2,_1>>,
   cute::Layout<_4>
 > {
@@ -363,9 +363,9 @@ struct LayoutAwareConvertImpl<
                 >& dst) {
 
     static_assert(cute::is_same_v<cutlass::int8_t, typename EngineIn::value_type> &&
-                  cute::is_same_v<cutlass::half_t, typename EngineOut::value_type>);
+                  cute::is_same_v<cutlass::bfloat16_t, typename EngineOut::value_type>);
     using SrcArray = cutlass::Array<cutlass::int8_t, 8>;
-    using DstArray = cutlass::Array<cutlass::half_t, 8>;
+    using DstArray = cutlass::Array<cutlass::bfloat16_t, 8>;
     using RegArray = cutlass::AlignedArray<uint32_t, 4, sizeof(DstArray)>;
 
     auto&& src_reg = cute::recast<uint32_t>(src)(0);
@@ -403,7 +403,7 @@ struct LayoutAwareConvertImpl<
 template <>
 struct LayoutAwareConvertImpl<
   cutlass::int8_t,
-  cutlass::bfloat16_t,
+  cutlass::half_t,
   cute::Layout<cute::Shape<_2,_2>, cute::Stride<_2,_1>>,
   cute::Layout<_4>
 > {
@@ -418,9 +418,9 @@ struct LayoutAwareConvertImpl<
                 >& dst) {
 
     static_assert(cute::is_same_v<cutlass::int8_t, typename EngineIn::value_type> &&
-                  cute::is_same_v<cutlass::bfloat16_t, typename EngineOut::value_type>);
+                  cute::is_same_v<cutlass::half_t, typename EngineOut::value_type>);
     using SrcArray = cutlass::Array<cutlass::int8_t, 8>;
-    using DstArray = cutlass::Array<cutlass::bfloat16_t, 8>;
+    using DstArray = cutlass::Array<cutlass::half_t, 8>;
     using RegArray = cutlass::AlignedArray<uint32_t, 4, sizeof(DstArray)>;
 
     auto&& src_reg = cute::recast<uint32_t>(src)(0);
